# docs/swagger.json
{
    "components": {
        "schemas": {
...
            "CategoryCreationRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 5
                    },
                    "type": {
                        "enum": [
                            "Item",
                            "Service"
                        ],
                        "type": "string"
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 250,
                        "minLength": 10
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "type"
                ]
            },
            "CategoryUpdateRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 5
                    },
                    "type": {
                        "enum": [
                            "Item",
                            "Service"
                        ],
                        "type": "string"
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 250,
                        "minLength": 10
                    }
                },
                "type": "object"
            },
...
            "ProfileUpdateRequest": {
                "properties": {
                    "firstName": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 2
                    },
                    "lastName": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 2
                    },
                    "phoneNumber": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string",
                        "maxLength": 150,
                        "minLength": 15
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 2
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 500,
                        "minLength": 50
                    },
                    "businessEmail": {
                        "format": "email",
                        "type": "string"
                    },
                    "logoUrl": {
                        "format": "url",
                        "type": "string"
                    },
                    "websiteUrl": {
                        "format": "url",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "SellerItemUpdateRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 75,
                        "minLength": 5
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 1000,
                        "minLength": 25
                    },
                    "price": {
                        "type": "number",
                        "minimum": 0
                    },
                    "categoryId": {
                        "pattern": "^[0-9a-fA-F]{24}$",
                        "type": "string"
                    },
                    "isAvailable": {
                        "type": "boolean"
                    },
                    "quantity": {
                        "type": "number",
                        "minimum": 0
                    },
                    "imageUrl": {
                        "format": "url",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
    "info": {
        "title": "sustainable-community-server API",
        "version": "1.0.0",
        "description": "Generated with `routing-controllers-openapi`"
    },
    "openapi": "3.0.0",
    "paths": {
...
        "/sellers/profile": {
            "put": {
                "operationId": "SellerController.updateProfile",
                "parameters": [
                    {
                        "in": "header",
                        "name": "auth",
                        "required": false,
                        "schema": {}
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProfileUpdateRequest"
                            }
                        }
                    },
                    "description": "ProfileUpdateRequest",
                    "required": false
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful response"
                    }
                },
                "summary": "Update seller profile",
                "tags": [
                    "Seller"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/sellers/{id}/categories/{categoryId}": {
            "put": {
                "operationId": "SellerController.updateCategory",
                "parameters": [
                    {
                        "in": "header",
                        "name": "auth",
                        "required": false,
                        "schema": {}
                    },
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    },
                    {
                        "in": "path",
                        "name": "categoryId",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful response"
                    }
                },
                "summary": "Update seller category",
                "tags": [
                    "Seller"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/items/{id}": {
            "get": {
                "operationId": "SellerItemController.getItem",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SellerItemResponse"
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "summary": "Get item by id",
                "tags": [
                    "Seller Item"
                ]
            },
            "put": {
                "operationId": "SellerItemController.updateItem",
                "parameters": [
                    {
                        "in": "header",
                        "name": "auth",
                        "required": false,
                        "schema": {}
                    },
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SellerItemUpdateRequest"
                            }
                        }
                    },
                    "description": "SellerItemUpdateRequest",
                    "required": false
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SellerItemResponse"
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "summary": "Update an item",
                "tags": [
                    "Seller Item"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "operationId": "SellerItemController.deleteItem",
                "parameters": [
                    {
                        "in": "header",
                        "name": "auth",
                        "required": false,
                        "schema": {}
                    },
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful response"
                    }
                },
                "summary": "Delete an item",
                "tags": [
                    "Seller Item"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/customers/wishlist/items/{itemId}": {
            "post": {
                "operationId": "CustomerController.addItemToWishlist",
                "parameters": [
                    {
                        "in": "path",
                        "name": "itemId",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful response"
                    }
                },
                "summary": "Add item to wishlist",
                "tags": [
                    "Customer"
                ]
            },
            "delete": {
                "operationId": "CustomerController.removeItemFromWishlist",
                "parameters": [
                    {
                        "in": "path",
                        "name": "itemId",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful response"
                    }
                },
                "summary": "Remove item from wishlist",
                "tags": [
                    "Customer"
                ]
            }
        },
...
        "/customers/profile": {
            "put": {
                "operationId": "CustomerController.updateProfile",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProfileUpdateRequest"
                            }
                        }
                    },
                    "description": "ProfileUpdateRequest",
                    "required": false
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful response"
                    }
                },
                "summary": "Update profile",
                "tags": [
                    "Customer"
                ]
            }
        },
        "/categories/{id}": {
            "get": {
                "operationId": "CategoryController.getOneCategory",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryResponse"
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "summary": "Get category by id",
                "tags": [
                    "Category"
                ]
            },
            "put": {
                "operationId": "CategoryController.updateCategory",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryUpdateRequest"
                            }
                        }
                    },
                    "description": "CategoryUpdateRequest",
                    "required": false
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryResponse"
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "summary": "Update a category",
                "tags": [
                    "Category"
                ]
            },
            "delete": {
                "operationId": "CategoryController.deleteCategory",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful response"
                    }
                },
                "summary": "Delete a category",
                "tags": [
                    "Category"
                ]
            }
        },
...
        "/categories/": {
            "get": {
                "operationId": "CategoryController.getListOfCategories",
                "parameters": [
                    {
                        "in": "query",
                        "name": "type",
                        "schema": {
                            "enum": [
                                "Item",
                                "Service"
                            ],
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/CategoryResponse"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "summary": "Get list of categories",
                "tags": [
                    "Category"
                ]
            },
            "post": {
                "operationId": "CategoryController.createCategory",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryCreationRequest"
                            }
                        }
                    },
                    "description": "CategoryCreationRequest",
                    "required": false
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryResponse"
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "summary": "Create a new category",
                "tags": [
                    "Category"
                ]
            }
        },
...
    }
}